Move contents of pdo.php to common.php
