+++ tests/CreatesApplication.php
@@ -19,7 +19,6 @@ trait CreatesApplication
         $app->make(Kernel::class)->bootstrap();
 
         require_once 'include/common.php';
-        require_once 'include/pdo.php';
 
         return $app;
     }
